@@ -1,12 +1,8 @@
 module github.com/muidea/magicOrm
 
-go 1.22
+go 1.24
 
-require (
-	github.com/go-sql-driver/mysql v1.8.1
-	github.com/muidea/magicCommon v1.3.122
-	github.com/stretchr/testify v1.8.4
-)
+require github.com/muidea/magicCommon v1.4.13
 
 require (
 	filippo.io/edwards25519 v1.1.0 // indirect
@@ -15,10 +11,13 @@ require (
 	github.com/go-playground/locales v0.14.1 // indirect
 	github.com/go-playground/universal-translator v0.18.1 // indirect
 	github.com/go-playground/validator/v10 v10.22.1 // indirect
+	github.com/go-sql-driver/mysql v1.8.1
+	github.com/google/go-cmp v0.7.0
 	github.com/google/uuid v1.6.0 // indirect
 	github.com/leodido/go-urn v1.4.0 // indirect
 	github.com/muidea/seelog v1.0.0 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/stretchr/testify v1.10.0
 	golang.org/x/crypto v0.28.0 // indirect
 	golang.org/x/net v0.30.0 // indirect
 	golang.org/x/sys v0.26.0 // indirect
