@@ -1,7 +1,9 @@
 package mysql
 
 import (
+	"bytes"
 	"fmt"
+	"strings"
 
 	cd "github.com/muidea/magicCommon/def"
 	"github.com/muidea/magicCommon/foundation/log"
@@ -9,11 +11,10 @@ import (
 	"github.com/muidea/magicOrm/model"
 )
 
-func (s *Builder) BuildCreateTable(vModel model.Model) (ret *ResultStack, err *cd.Result) {
+func (s *Builder) BuildCreateTable(vModel model.Model) (ret *ResultStack, err *cd.Error) {
 	createSQL := ""
 	for _, field := range vModel.GetFields() {
-		fType := field.GetType()
-		if !fType.IsBasic() {
+		if !model.IsBasicField(field) {
 			continue
 		}
 
@@ -39,12 +40,12 @@ func (s *Builder) BuildCreateTable(vModel model.Model) (ret *ResultStack, err *c
 		log.Infof("[SQL] create: %s", createSQL)
 	}
 
-	ret = NewResult(createSQL, nil)
+	ret = NewError(createSQL, nil)
 	return
 }
 
 // BuildCreateRelationTable Build CreateRelation Schema
-func (s *Builder) BuildCreateRelationTable(vModel model.Model, vField model.Field, rModel model.Model) (ret *ResultStack, err *cd.Result) {
+func (s *Builder) BuildCreateRelationTable(vModel model.Model, vField model.Field) (ret *ResultStack, err *cd.Error) {
 	lPKField := vModel.GetPrimaryField()
 	lPKType, lPKErr := getTypeDeclare(lPKField.GetType(), lPKField.GetSpec())
 	if lPKErr != nil {
@@ -53,10 +54,17 @@ func (s *Builder) BuildCreateRelationTable(vModel model.Model, vField model.Fiel
 		return
 	}
 
-	relationTableName, relationErr := s.buildCodec.ConstructRelationTableName(vModel, vField, rModel)
+	relationTableName, relationErr := s.buildCodec.ConstructRelationTableName(vModel, vField)
 	if relationErr != nil {
 		err = relationErr
 		log.Errorf("BuildCreateRelationTable %s failed, s.buildCodec.ConstructRelationTableName error:%s", vField.GetName(), err.Error())
+		return
+	}
+
+	rModel, rErr := s.modelProvider.GetTypeModel(vField.GetType().Elem())
+	if rErr != nil {
+		err = rErr
+		log.Errorf("BuildCreateRelationTable %s failed, s.modelProvider.GetTypeModel error:%s", vField.GetName(), err.Error())
 		return
 	}
 
@@ -74,25 +82,114 @@ func (s *Builder) BuildCreateRelationTable(vModel model.Model, vField model.Fiel
 		log.Infof("[SQL] create relation: %s", createRelationSQL)
 	}
 
-	ret = NewResult(createRelationSQL, nil)
+	ret = NewError(createRelationSQL, nil)
 	return
 }
 
-func (s *Builder) declareFieldInfo(vField model.Field) (ret string, err *cd.Result) {
-	autoIncrement := ""
-	fSpec := vField.GetSpec()
-	if fSpec != nil && model.IsAutoIncrement(fSpec.GetValueDeclare()) {
-		autoIncrement = "AUTO_INCREMENT"
-	}
+// declareFieldInfo declare field info
+// 根据字段类型和字段特性生成字段定义
+// 类似以下信息
+// `id` int(11) NOT NULL AUTO_INCREMENT
+// `i8` tinyint(4) DEFAULT '100',
+func (s *Builder) declareFieldInfo(vField model.Field) (ret string, err *cd.Error) {
+	strBuffer := bytes.NewBufferString("")
+	// Write field name
+	strBuffer.WriteString("`")
+	strBuffer.WriteString(vField.GetName())
+	strBuffer.WriteString("`")
 
-	allowNull := "NOT NULL"
+	// Write field type
 	typeVal, typeErr := getTypeDeclare(vField.GetType(), vField.GetSpec())
 	if typeErr != nil {
 		err = typeErr
 		log.Errorf("declareFieldInfo failed, getTypeDeclare error:%s", err.Error())
 		return
 	}
+	strBuffer.WriteString(" ")
+	strBuffer.WriteString(typeVal)
 
-	ret = fmt.Sprintf("`%s` %s %s %s", vField.GetName(), typeVal, allowNull, autoIncrement)
+	// Write NULL constraint
+	if !vField.GetType().IsPtrType() {
+		strBuffer.WriteString(" NOT NULL")
+	}
+
+	// Write default value if exists
+	fSpec := vField.GetSpec()
+	defaultValue, defaultErr := s.validDefaultValue(vField.GetType(), fSpec)
+	if defaultErr != nil {
+		err = defaultErr
+		log.Errorf("declareFieldInfo failed, validDefaultValue error:%s", err.Error())
+		return
+	}
+	// Write auto increment if needed
+	autoIncVal, autoIncErr := s.validAutoIncrement(vField.GetType(), vField.GetSpec())
+	if autoIncErr != nil {
+		err = autoIncErr
+		log.Errorf("declareFieldInfo failed, validAutoIncrement error:%s", err.Error())
+		return
+	}
+
+	if !autoIncVal && defaultValue != "''" && defaultValue != "" {
+		strBuffer.WriteString(" DEFAULT ")
+		strBuffer.WriteString(defaultValue)
+	}
+
+	if autoIncVal {
+		strBuffer.WriteString(" AUTO_INCREMENT")
+	}
+
+	ret = strBuffer.String()
+	return
+}
+
+func (s *Builder) validDefaultValue(vType model.Type, vSpec model.Spec) (ret string, err *cd.Error) {
+	if !model.IsBasic(vType) || vType.IsPtrType() || vType.GetValue().IsSliceType() || vType.Elem().GetValue().IsStringValueType() {
+		// 非基础类型和切片类型不需要设置默认值
+		// 指针类型不需要设置默认值
+		// 字符串类型不需要设置默认值，这里返回空
+		return
+	}
+
+	var defaultValue, defaultValueDeclare any
+	if vSpec != nil {
+		defaultValueDeclare = vSpec.GetDefaultValue()
+	}
+	if defaultValueDeclare != nil {
+		switch val := defaultValueDeclare.(type) {
+		case string:
+			if strings.Contains(val, "$referenceValue.") {
+				vTypeDefaultVal, _ := vType.Interface(nil)
+				defaultValue = vTypeDefaultVal.Get()
+			} else {
+				defaultValue = val
+			}
+		default:
+			defaultValue = val
+		}
+	} else {
+		vTypeDefaultVal, _ := vType.Interface(nil)
+		defaultValue = vTypeDefaultVal.Get()
+	}
+
+	switch vType.Elem().GetValue() {
+	case model.TypeBooleanValue:
+		if defaultValue.(bool) {
+			ret = "'1'"
+		} else {
+			ret = "'0'"
+		}
+	default:
+		ret = fmt.Sprintf("'%v'", defaultValue)
+	}
+
+	return
+}
+
+func (s *Builder) validAutoIncrement(vType model.Type, vSpec model.Spec) (ret bool, err *cd.Error) {
+	if vSpec == nil || !vSpec.IsPrimaryKey() || !vType.GetValue().IsNumberValueType() {
+		return
+	}
+
+	ret = model.IsAutoIncrementDeclare(vSpec.GetValueDeclare())
 	return
 }
