@@ -10,20 +10,20 @@ import (
 )
 
 // BuildDropTable  BuildDropSchema
-func (s *Builder) BuildDropTable(vModel model.Model) (ret *ResultStack, err *cd.Result) {
+func (s *Builder) BuildDropTable(vModel model.Model) (ret *ResultStack, err *cd.Error) {
 	dropSQL := fmt.Sprintf("DROP TABLE IF EXISTS `%s`", s.buildCodec.ConstructModelTableName(vModel))
 	//log.Print(dropSQL)
 	if traceSQL() {
 		log.Infof("[SQL] drop: %s", dropSQL)
 	}
 
-	ret = NewResult(dropSQL, nil)
+	ret = NewError(dropSQL, nil)
 	return
 }
 
 // BuildDropRelationTable Build DropRelation Schema
-func (s *Builder) BuildDropRelationTable(vModel model.Model, vField model.Field, rModel model.Model) (ret *ResultStack, err *cd.Result) {
-	relationTableName, relationErr := s.buildCodec.ConstructRelationTableName(vModel, vField, rModel)
+func (s *Builder) BuildDropRelationTable(vModel model.Model, vField model.Field) (ret *ResultStack, err *cd.Error) {
+	relationTableName, relationErr := s.buildCodec.ConstructRelationTableName(vModel, vField)
 	if relationErr != nil {
 		err = relationErr
 		log.Errorf("BuildDeleteRelation %s failed, s.buildCodec.ConstructRelationTableName error:%s", vField.GetName(), err.Error())
@@ -36,6 +36,6 @@ func (s *Builder) BuildDropRelationTable(vModel model.Model, vField model.Field,
 		log.Infof("[SQL] drop relation: %s", dropRelationSQL)
 	}
 
-	ret = NewResult(dropRelationSQL, nil)
+	ret = NewError(dropRelationSQL, nil)
 	return
 }
