@@ -26,7 +26,7 @@ var mode = 1
 
 var finishFlag = false
 
-type funcPtr func(executor *Executor) *cd.Result
+type funcPtr func(executor *Executor) *cd.Error
 
 func TestMasterSlaver(t *testing.T) {
 	pool := NewPool()
@@ -100,7 +100,7 @@ func testDML(wg *sync.WaitGroup, pool *Pool) {
 	pickExecutor(pool, wg, createSchema)
 
 	wg.Add(1)
-	pickExecutor(pool, wg, func(executor *Executor) (err *cd.Result) {
+	pickExecutor(pool, wg, func(executor *Executor) (err *cd.Error) {
 		bVal, bErr := checkSchema(executor)
 		if bErr != nil {
 			log.Errorf("checkSchema failed, error:%s", bErr.Error())
@@ -176,24 +176,24 @@ func pickExecutor(pool *Pool, wg *sync.WaitGroup, fPtr funcPtr) {
 	wg.Done()
 }
 
-func createSchema(executor *Executor) (err *cd.Result) {
+func createSchema(executor *Executor) (err *cd.Error) {
 	sql := "CREATE TABLE `Unit001` (\n\t`id` INT NOT NULL AUTO_INCREMENT,\n\t`i8` TINYINT NOT NULL ,\n\t`i16` SMALLINT NOT NULL ,\n\t`i32` INT NOT NULL ,\n\t`i64` BIGINT NOT NULL ,\n\t`name` TEXT NOT NULL ,\n\t`value` FLOAT NOT NULL ,\n\t`f64` DOUBLE NOT NULL ,\n\t`ts` DATETIME NOT NULL ,\n\t`flag` TINYINT NOT NULL ,\n\t`iArray` TEXT NOT NULL ,\n\t`fArray` TEXT NOT NULL ,\n\t`strArray` TEXT NOT NULL ,\n\tPRIMARY KEY (`id`)\n)"
 	_, _, err = executor.Execute(sql)
 	return
 }
 
-func dropSchema(executor *Executor) (err *cd.Result) {
+func dropSchema(executor *Executor) (err *cd.Error) {
 	sql := "DROP TABLE IF EXISTS `Unit001`"
 	_, _, err = executor.Execute(sql)
 	return
 }
 
-func checkSchema(executor *Executor) (ret bool, err *cd.Result) {
+func checkSchema(executor *Executor) (ret bool, err *cd.Error) {
 	ret, err = executor.CheckTableExist(`Unit001`)
 	return
 }
 
-func insertValue(executor *Executor) (err *cd.Result) {
+func insertValue(executor *Executor) (err *cd.Error) {
 	sql := "INSERT INTO `Unit001` (`i8`,`i16`,`i32`,`i64`,`name`,`value`,`f64`,`ts`,`flag`,`iArray`,`fArray`,`strArray`) VALUES (8,1600,323200,78962222222,'Hello world',12.345600128173828,12.45678,'2018-01-02 15:04:05',1,'12','12.34','abcdef')"
 	idx := 0
 	for idx < itemSize/threadSize {
@@ -203,37 +203,37 @@ func insertValue(executor *Executor) (err *cd.Result) {
 	return
 }
 
-func alterSchemaAdd(executor *Executor) (err *cd.Result) {
+func alterSchemaAdd(executor *Executor) (err *cd.Error) {
 	sql := "ALTER TABLE `Unit001` ADD dVal DATE"
 	_, _, err = executor.Execute(sql)
 	return
 }
 
-func alterSchemaDrop(executor *Executor) (err *cd.Result) {
+func alterSchemaDrop(executor *Executor) (err *cd.Error) {
 	sql := "ALTER TABLE `Unit001` DROP dVal"
 	_, _, err = executor.Execute(sql)
 	return
 }
 
-func alterSchemaOlnDDLAdd(executor *Executor) (err *cd.Result) {
+func alterSchemaOlnDDLAdd(executor *Executor) (err *cd.Error) {
 	sql := "ALTER TABLE `Unit001` ADD dVal2 DATE, ALGORITHM=DEFAULT, LOCK=NONE"
 	_, _, err = executor.Execute(sql)
 	return
 }
 
-func alterSchemaOlnDDLDrop(executor *Executor) (err *cd.Result) {
+func alterSchemaOlnDDLDrop(executor *Executor) (err *cd.Error) {
 	sql := "ALTER TABLE `Unit001` DROP dVal2, ALGORITHM=DEFAULT, LOCK=NONE"
 	_, _, err = executor.Execute(sql)
 	return
 }
 
-func createSchemaDDL(executor *Executor) (err *cd.Result) {
+func createSchemaDDL(executor *Executor) (err *cd.Error) {
 	sql := "CREATE TABLE `Unit002` (\n\t`id` INT NOT NULL AUTO_INCREMENT,\n\t`i8` TINYINT NOT NULL ,\n\t`i16` SMALLINT NOT NULL ,\n\t`i32` INT NOT NULL ,\n\t`i64` BIGINT NOT NULL ,\n\t`name` TEXT NOT NULL ,\n\t`value` FLOAT NOT NULL ,\n\t`f64` DOUBLE NOT NULL ,\n\t`ts` DATETIME NOT NULL ,\n\t`flag` TINYINT NOT NULL ,\n\t`iArray` TEXT NOT NULL ,\n\t`fArray` TEXT NOT NULL ,\n\t`strArray` TEXT NOT NULL ,\n\tPRIMARY KEY (`id`)\n)"
 	_, _, err = executor.Execute(sql)
 	return
 }
 
-func dropSchemaDDL(executor *Executor) (err *cd.Result) {
+func dropSchemaDDL(executor *Executor) (err *cd.Error) {
 	sql := "DROP TABLE IF EXISTS `Unit002`"
 	_, _, err = executor.Execute(sql)
 	return
