@@ -19,7 +19,7 @@ func traceSQL() bool {
 	return false
 }
 
-func getTypeDeclare(fType model.Type, fSpec model.Spec) (ret string, err *cd.Result) {
+func getTypeDeclare(fType model.Type, fSpec model.Spec) (ret string, err *cd.Error) {
 	switch fType.GetValue() {
 	case model.TypeStringValue:
 		if fSpec.IsPrimaryKey() {
@@ -44,7 +44,7 @@ func getTypeDeclare(fType model.Type, fSpec model.Spec) (ret string, err *cd.Res
 	case model.TypeSliceValue:
 		ret = "TEXT"
 	default:
-		err = cd.NewResult(cd.UnExpected, fmt.Sprintf("no support field type, type:%v", fType.GetPkgKey()))
+		err = cd.NewError(cd.Unexpected, fmt.Sprintf("no support field type, type:%v", fType.GetPkgKey()))
 	}
 
 	if err != nil {
@@ -54,8 +54,7 @@ func getTypeDeclare(fType model.Type, fSpec model.Spec) (ret string, err *cd.Res
 	return
 }
 
-func getFieldPlaceHolder(field model.Field) (ret interface{}, err *cd.Result) {
-	fType := field.GetType()
+func getFieldPlaceHolder(fType model.Type) (ret interface{}, err *cd.Error) {
 	switch fType.GetValue() {
 	case model.TypeStringValue, model.TypeDateTimeValue:
 		val := ""
@@ -94,14 +93,14 @@ func getFieldPlaceHolder(field model.Field) (ret interface{}, err *cd.Result) {
 		val := 0.0000
 		ret = &val
 	case model.TypeSliceValue:
-		if fType.IsBasic() {
+		if model.IsBasic(fType.Elem()) {
 			val := ""
 			ret = &val
 		} else {
-			err = cd.NewResult(cd.UnExpected, fmt.Sprintf("no support fileType, name:%s, type:%v", field.GetName(), fType.GetPkgKey()))
+			err = cd.NewError(cd.Unexpected, fmt.Sprintf("no support fileType, type:%v", fType.GetPkgKey()))
 		}
 	default:
-		err = cd.NewResult(cd.UnExpected, fmt.Sprintf("no support fileType, name:%s, type:%v", field.GetName(), fType.GetPkgKey()))
+		err = cd.NewError(cd.Unexpected, fmt.Sprintf("no support fileType, type:%v", fType.GetPkgKey()))
 	}
 
 	if err != nil {
