@@ -29,20 +29,14 @@ func NewInsertRunner(
 	}
 }
 
-func (s *InsertRunner) insertHost(vModel model.Model) (err *cd.Result) {
+func (s *InsertRunner) insertHost(vModel model.Model) (err *cd.Error) {
 	autoIncrementFlag := false
 	for _, field := range vModel.GetFields() {
-		if !field.IsBasic() {
+		if !model.IsBasicField(field) {
 			continue
 		}
 
-		fSpec := field.GetSpec()
-		fValue := field.GetValue()
-		if !fValue.IsValid() {
-			fValue = s.modelProvider.GetNewValue(fSpec.GetValueDeclare())
-			field.SetValue(fValue)
-		}
-		if fSpec.GetValueDeclare() == model.AutoIncrement {
+		if field.GetSpec().GetValueDeclare() == model.AutoIncrement {
 			autoIncrementFlag = true
 		}
 	}
@@ -55,20 +49,23 @@ func (s *InsertRunner) insertHost(vModel model.Model) (err *cd.Result) {
 	}
 
 	if pkVal != nil && autoIncrementFlag {
-		pkField := vModel.GetPrimaryField()
-		tVal, tErr := s.modelProvider.DecodeValue(pkVal, pkField.GetType())
-		if tErr != nil {
-			err = tErr
-			log.Errorf("insertHost failed, s.modelProvider.DecodeValue error:%s", err.Error())
+		pkFiled := vModel.GetPrimaryField()
+		vVal, vErr := s.modelCodec.ExtractBasicFieldValue(pkFiled, pkVal)
+		if vErr != nil {
+			err = vErr
+			log.Errorf("insertHost failed, s.modelCodec.ExtractFieldValue error:%s", err.Error())
 			return
 		}
-
-		pkField.SetValue(tVal)
+		err = pkFiled.SetValue(vVal)
+		if err != nil {
+			log.Errorf("insertHost failed, s.modelCodec.ExtractFieldValue error:%s", err.Error())
+			return
+		}
 	}
 	return
 }
 
-func (s *InsertRunner) innerHost(vModel model.Model) (ret model.RawVal, err *cd.Result) {
+func (s *InsertRunner) innerHost(vModel model.Model) (ret any, err *cd.Error) {
 	insertResult, insertErr := s.hBuilder.BuildInsert(vModel)
 	if insertErr != nil {
 		err = insertErr
@@ -83,55 +80,46 @@ func (s *InsertRunner) innerHost(vModel model.Model) (ret model.RawVal, err *cd.
 		return
 	}
 
-	ret = model.NewRawVal(id)
+	ret = id
 	return
 }
 
-func (s *InsertRunner) insertRelation(vModel model.Model, vField model.Field, rModel model.Model) (err *cd.Result) {
-	fValue := vField.GetValue()
-	if fValue.IsZero() {
-		return
-	}
-
-	if vField.IsSlice() {
-		rValue, rErr := s.insertSliceRelation(vModel, vField, rModel)
+func (s *InsertRunner) insertRelation(vModel model.Model, vField model.Field) (err *cd.Error) {
+	if model.IsSliceField(vField) {
+		rErr := s.insertSliceRelation(vModel, vField)
 		if rErr != nil {
 			err = rErr
 			log.Errorf("insertRelation failed, s.insertSliceRelation error:%s", err.Error())
 			return
 		}
-
-		if rValue != nil && !rValue.IsZero() {
-			vField.SetValue(rValue)
-		}
 		return
 	}
 
-	rValue, rErr := s.insertSingleRelation(vModel, vField, rModel)
+	rErr := s.insertSingleRelation(vModel, vField)
 	if rErr != nil {
 		err = rErr
 		log.Errorf("insertRelation failed, s.insertSingleRelation error:%s", err.Error())
 		return
 	}
-
-	if rValue != nil && !rValue.IsZero() {
-		vField.SetValue(rValue)
-	}
 	return
 }
 
-func (s *InsertRunner) insertSingleRelation(vModel model.Model, vField model.Field, rModel model.Model) (ret model.Value, err *cd.Result) {
-	fType := vField.GetType()
-	fValue := vField.GetValue()
-	var rErr *cd.Result
-	rModel, rErr = s.modelProvider.SetModelValue(rModel, fValue)
+func (s *InsertRunner) insertSingleRelation(vModel model.Model, vField model.Field) (err *cd.Error) {
+	elemType := vField.GetType().Elem()
+	rModel, rErr := s.modelProvider.GetTypeModel(elemType)
+	if rErr != nil {
+		err = rErr
+		log.Errorf("insertSingleRelation failed, s.modelProvider.GetTypeModel error:%s", err.Error())
+		return
+	}
+	rModel, rErr = s.modelProvider.SetModelValue(rModel, vField.GetValue())
 	if rErr != nil {
 		err = rErr
 		log.Errorf("insertSingleRelation failed, s.modelProvider.SetModelValue error:%s", err.Error())
 		return
 	}
 
-	if !vField.IsPtrType() {
+	if !model.IsPtrField(vField) {
 		rInsertRunner := NewInsertRunner(rModel, s.executor, s.modelProvider, s.modelCodec)
 		rModel, rErr = rInsertRunner.Insert()
 		if rErr != nil {
@@ -154,44 +142,33 @@ func (s *InsertRunner) insertSingleRelation(vModel model.Model, vField model.Fie
 		return
 	}
 
-	entityVal, entityErr := s.modelProvider.GetEntityValue(rModel.Interface(fType.IsPtrType(), model.LiteView))
-	if entityErr != nil {
-		err = entityErr
-		log.Errorf("insertSingleRelation failed, s.modelProvider.GetEntityValue error:%s", err.Error())
-		return
-	}
-
-	ret = entityVal
+	vField.SetValue(rModel.Interface(model.IsPtrField(vField)))
 	return
 }
 
-func (s *InsertRunner) insertSliceRelation(vModel model.Model, vField model.Field, rModel model.Model) (ret model.Value, err *cd.Result) {
-	fValue := vField.GetValue()
-	fType := vField.GetType()
-	rvValue, _ := fType.Interface(nil)
-	fSliceValue, fSliceErr := s.modelProvider.ElemDependValue(fValue.Interface())
-	if fSliceErr != nil {
-		err = fSliceErr
-		log.Errorf("insertSliceRelation failed, s.modelProvider.ElemDependValue error:%s", err.Error())
-		return
-	}
-
-	var rErr *cd.Result
+func (s *InsertRunner) insertSliceRelation(vModel model.Model, vField model.Field) (err *cd.Error) {
+	fSliceValue := vField.GetSliceValue()
 	for _, fVal := range fSliceValue {
-		rModel, rErr = s.modelProvider.SetModelValue(rModel.Copy(true), fVal)
+		elemType := vField.GetType().Elem()
+		rModel, rErr := s.modelProvider.GetTypeModel(elemType)
 		if rErr != nil {
 			err = rErr
-			log.Errorf("insertSliceRelation failed, s.modelProvider.GetValueModel error:%s", err.Error())
+			log.Errorf("insertSliceRelation failed, model:%s, filed name:%s, s.modelProvider.GetTypeModel error:%s", vModel.GetPkgKey(), vField.GetName(), err.Error())
+			return
+		}
+		rModel, rErr = s.modelProvider.SetModelValue(rModel, fVal)
+		if rErr != nil {
+			err = rErr
+			log.Errorf("insertSliceRelation failed, model:%s, filed name:%s, s.modelProvider.SetModelValue error:%s", vModel.GetPkgKey(), vField.GetName(), err.Error())
 			return
 		}
 
-		elemType := fType.Elem()
 		if !elemType.IsPtrType() {
 			rInsertRunner := NewInsertRunner(rModel, s.executor, s.modelProvider, s.modelCodec)
 			rModel, rErr = rInsertRunner.Insert()
 			if rErr != nil {
 				err = rErr
-				log.Errorf("insertSliceRelation failed, s.insertSingle error:%s", err.Error())
+				log.Errorf("insertSliceRelation failed, model:%s, filed name:%s, s.insertSingle error:%s", vModel.GetPkgKey(), vField.GetName(), err.Error())
 				return
 			}
 		}
@@ -199,35 +176,27 @@ func (s *InsertRunner) insertSliceRelation(vModel model.Model, vField model.Fiel
 		relationResult, relationErr := s.hBuilder.BuildInsertRelation(vModel, vField, rModel)
 		if relationErr != nil {
 			err = relationErr
-			log.Errorf("insertSliceRelation failed, s.hBuilder.BuildInsertRelation error:%s", err.Error())
+			log.Errorf("insertSliceRelation failed, model:%s, filed name:%s, s.hBuilder.BuildInsertRelation error:%s", vModel.GetPkgKey(), vField.GetName(), err.Error())
 			return
 		}
 
 		_, _, err = s.executor.Execute(relationResult.SQL(), relationResult.Args()...)
 		if err != nil {
-			log.Errorf("insertSliceRelation failed, s.executor.Execute error:%s", err.Error())
+			log.Errorf("insertSliceRelation failed, model:%s, filed name:%s, s.executor.Execute error:%s", vModel.GetPkgKey(), vField.GetName(), err.Error())
 			return
 		}
 
-		entityVal, entityErr := s.modelProvider.GetEntityValue(rModel.Interface(elemType.IsPtrType(), model.LiteView))
-		if entityErr != nil {
-			err = entityErr
-			log.Errorf("insertSliceRelation failed, s.modelProvider.GetEntityValue error:%s", err.Error())
-			return
-		}
-
-		rvValue, err = s.modelProvider.AppendSliceValue(rvValue, entityVal)
+		// 这里只需要直接更新值就可以
+		err = fVal.Set(rModel.Interface(elemType.IsPtrType()))
 		if err != nil {
-			log.Errorf("insertSliceRelation failed, s.modelProvider.AppendSliceValue error:%s", err.Error())
+			log.Errorf("insertSliceRelation failed, model:%s, filed name:%s, fVal.Set error:%s", vModel.GetPkgKey(), vField.GetName(), err.Error())
 			return
 		}
 	}
-
-	ret = rvValue
 	return
 }
 
-func (s *InsertRunner) Insert() (ret model.Model, err *cd.Result) {
+func (s *InsertRunner) Insert() (ret model.Model, err *cd.Error) {
 	err = s.insertHost(s.vModel)
 	if err != nil {
 		log.Errorf("Insert failed, s.insertSingle error:%s", err.Error())
@@ -235,18 +204,11 @@ func (s *InsertRunner) Insert() (ret model.Model, err *cd.Result) {
 	}
 
 	for _, field := range s.vModel.GetFields() {
-		if field.IsBasic() {
+		if model.IsBasicField(field) || !model.IsValidField(field) {
 			continue
 		}
 
-		rModel, rErr := s.modelProvider.GetTypeModel(field.GetType())
-		if rErr != nil {
-			err = rErr
-			log.Errorf("Insert failed, s.modelProvider.GetTypeModel error:%s", err.Error())
-			return
-		}
-
-		err = s.insertRelation(s.vModel, field, rModel)
+		err = s.insertRelation(s.vModel, field)
 		if err != nil {
 			log.Errorf("Insert failed, s.insertRelation error:%s", err.Error())
 			return
@@ -257,9 +219,9 @@ func (s *InsertRunner) Insert() (ret model.Model, err *cd.Result) {
 	return
 }
 
-func (s *impl) Insert(vModel model.Model) (ret model.Model, err *cd.Result) {
+func (s *impl) Insert(vModel model.Model) (ret model.Model, err *cd.Error) {
 	if vModel == nil {
-		err = cd.NewResult(cd.IllegalParam, "illegal model value")
+		err = cd.NewError(cd.IllegalParam, "illegal model value")
 		return
 	}
 
