@@ -2,6 +2,7 @@ package orm
 
 import (
 	"fmt"
+	"reflect"
 
 	cd "github.com/muidea/magicCommon/def"
 	"github.com/muidea/magicCommon/foundation/log"
@@ -32,7 +33,7 @@ func NewQueryRunner(
 	}
 }
 
-func (s *QueryRunner) innerQuery(vModel model.Model, filter model.Filter) (ret resultItemsList, err *cd.Result) {
+func (s *QueryRunner) innerQuery(vModel model.Model, filter model.Filter) (ret resultItemsList, err *cd.Error) {
 	queryResult, queryErr := s.hBuilder.BuildQuery(vModel, filter)
 	if queryErr != nil {
 		err = queryErr
@@ -52,11 +53,7 @@ func (s *QueryRunner) innerQuery(vModel model.Model, filter model.Filter) (ret r
 		itemValues, itemErr := s.hBuilder.BuildQueryPlaceHolder(vModel)
 		if itemErr != nil {
 			err = itemErr
-			if err.Fail() {
-				log.Errorf("innerQuery failed, getModelFieldsPlaceHolder error:%s", err.Error())
-			} else if err.Warn() {
-				log.Warnf("innerQuery failed, getModelFieldsPlaceHolder error:%s", err.Error())
-			}
+			log.Errorf("innerQuery failed, getModelFieldsPlaceHolder error:%s", err.Error())
 			return
 		}
 
@@ -73,48 +70,28 @@ func (s *QueryRunner) innerQuery(vModel model.Model, filter model.Filter) (ret r
 	return
 }
 
-func (s *QueryRunner) innerAssign(vModel model.Model, queryVal resultItems, deepLevel int) (ret model.Model, err *cd.Result) {
+func (s *QueryRunner) innerAssign(vModel model.Model, queryVal resultItems, deepLevel int) (ret model.Model, err *cd.Error) {
 	offset := 0
-	qModel := vModel.Copy(false)
+	qModel := vModel.Copy(model.OriginView)
 	for _, field := range qModel.GetFields() {
-		fType := field.GetType()
-		fValue := field.GetValue()
-		if !fType.IsBasic() || !fValue.IsValid() {
+		if !model.IsBasicField(field) || !model.IsValidField(field) {
 			continue
 		}
-
 		err = s.assignBasicField(field, queryVal[offset])
 		if err != nil {
-			if err.Fail() {
-				log.Errorf("innerAssign field:%s failed, s.assignBasicField error:%v", field.GetName(), err.Error())
-			} else if err.Warn() {
-				log.Warnf("innerAssign field:%s failed, s.assignBasicField error:%v", field.GetName(), err.Error())
-			}
+			log.Errorf("innerAssign field:%s failed, s.assignBasicField error:%v", field.GetName(), err.Error())
 			return
 		}
 		offset++
 	}
 
 	for _, field := range qModel.GetFields() {
-		fType := field.GetType()
-		fValue := field.GetValue()
-		if fType.IsBasic() || !fValue.IsValid() {
+		if model.IsBasicField(field) || !model.IsValidField(field) {
 			continue
 		}
-
-		rModel, rErr := s.modelProvider.GetTypeModel(field.GetType())
-		if rErr != nil {
-			err = rErr
-			return
-		}
-
-		err = s.assignModelField(qModel, field, rModel, deepLevel)
+		err = s.assignModelField(qModel, field, deepLevel)
 		if err != nil {
-			if err.Fail() {
-				log.Errorf("innerAssign field:%s failed, s.assignModelField error:%v", field.GetName(), err.Error())
-			} else if err.Warn() {
-				log.Warnf("innerAssign field:%s failed, s.assignModelField error:%v", field.GetName(), err.Error())
-			}
+			log.Errorf("innerAssign field:%s failed, s.assignModelField error:%v", field.GetName(), err.Error())
 			return
 		}
 	}
@@ -123,24 +100,19 @@ func (s *QueryRunner) innerAssign(vModel model.Model, queryVal resultItems, deep
 	return
 }
 
-func (s *QueryRunner) assignModelField(vModel model.Model, vField model.Field, rModel model.Model, deepLevel int) (err *cd.Result) {
-	vVal, vErr := s.queryRelation(vModel, vField, rModel, deepLevel)
+func (s *QueryRunner) assignModelField(vModel model.Model, vField model.Field, deepLevel int) (err *cd.Error) {
+	vErr := s.queryRelation(vModel, vField, deepLevel)
 	if vErr != nil {
 		err = vErr
-		if err.Fail() {
-			log.Errorf("assignModelField field:%s failed, s.queryRelation error:%v", vField.GetName(), err.Error())
-		} else if err.Warn() {
-			log.Warnf("assignModelFiel field:%sd failed, s.queryRelation error:%v", vField.GetName(), err.Error())
-		}
+		log.Errorf("assignModelField field:%s failed, s.queryRelation error:%v", vField.GetName(), err.Error())
 		return
 	}
 
-	vField.SetValue(vVal)
 	return
 }
 
-func (s *QueryRunner) assignBasicField(vField model.Field, val any) (err *cd.Result) {
-	fVal, fErr := s.modelCodec.ExtractFiledValue(vField, model.NewRawVal(val))
+func (s *QueryRunner) assignBasicField(vField model.Field, val any) (err *cd.Error) {
+	fVal, fErr := s.modelCodec.ExtractBasicFieldValue(vField, val)
 	if fErr != nil {
 		err = fErr
 		log.Errorf("assignBasicField field:%s failed, s.modelProvider.DecodeValue error:%v", vField.GetName(), err.Error())
@@ -151,148 +123,76 @@ func (s *QueryRunner) assignBasicField(vField model.Field, val any) (err *cd.Res
 	return
 }
 
-func (s *QueryRunner) queryRelation(vModel model.Model, vField model.Field, rModel model.Model, deepLevel int) (ret model.Value, err *cd.Result) {
+func (s *QueryRunner) queryRelation(vModel model.Model, vField model.Field, deepLevel int) (err *cd.Error) {
 	if deepLevel > maxDeepLevel {
-		fieldType := vField.GetType()
-		ret, _ = fieldType.Interface(nil)
 		return
 	}
 
-	if vField.IsSlice() {
-		ret, err = s.querySliceRelation(vModel, vField, rModel, deepLevel)
+	if model.IsSliceField(vField) {
+		err = s.querySliceRelation(vModel, vField, deepLevel)
 		if err != nil {
-			if err.Fail() {
-				log.Errorf("queryRelation field:%s failed, s.querySliceRelation error:%v", vField.GetName(), err.Error())
-			} else if err.Warn() {
-				log.Warnf("queryRelation field:%s failed, s.querySliceRelation error:%v", vField.GetName(), err.Error())
-			}
+			log.Errorf("queryRelation field:%s failed, s.querySliceRelation error:%v", vField.GetName(), err.Error())
 		}
 		return
 	}
 
-	ret, err = s.querySingleRelation(vModel, vField, rModel, deepLevel)
+	err = s.querySingleRelation(vModel, vField, deepLevel)
 	if err != nil {
-		if err.Fail() {
-			log.Errorf("queryRelation field:%s failed, s.querySingleRelation error:%v", vField.GetName(), err.Error())
-		} else if err.Warn() {
-			log.Warnf("queryRelation field:%s failed, s.querySingleRelation error:%v", vField.GetName(), err.Error())
-		}
+		log.Errorf("queryRelation field:%s failed, s.querySingleRelation error:%v", vField.GetName(), err.Error())
 	}
 	return
 }
 
-func (s *QueryRunner) querySingleRelation(vModel model.Model, vField model.Field, rModel model.Model, deepLevel int) (ret model.Value, err *cd.Result) {
-	valueList, valueErr := s.innerQueryRelationKeys(vModel, vField, rModel)
+func (s *QueryRunner) querySingleRelation(vModel model.Model, vField model.Field, deepLevel int) (err *cd.Error) {
+	valueList, valueErr := s.innerQueryRelationKeys(vModel, vField)
 	if valueErr != nil {
 		err = valueErr
-		if err.Fail() {
-			log.Errorf("querySingleRelation failed, s.innerQueryRelationKeys error:%v", err.Error())
-		} else if err.Warn() {
-			log.Warnf("querySingleRelation failed, s.innerQueryRelationKeys error:%v", err.Error())
-		}
+		log.Errorf("querySingleRelation failed, s.innerQueryRelationKeys error:%v", err.Error())
 		return
 	}
 
 	vType := vField.GetType()
 	valueSize := len(valueList)
 	if valueSize == 0 {
-		ret, _ = vType.Interface(nil)
 		if vType.IsPtrType() {
 			return
 		}
-
 		log.Warnf("query relation failed, field name:%s", vField.GetName())
 		return
 	}
 
-	rvModel, rvErr := s.innerQueryRelationSingleModel(valueList[0], rModel, deepLevel)
+	rvErr := s.innerQueryRelationSingleModel(valueList[0], vField, deepLevel)
 	if rvErr != nil {
-		if rvErr.Fail() {
-			log.Errorf("querySingleRelation failed, s.innerQueryRelationSingleModel error:%v", rvErr.Error())
-		} else if rvErr.Warn() {
-			log.Warnf("querySingleRelation failed, s.innerQueryRelationSingleModel error:%v", rvErr.Error())
-		}
-
+		log.Errorf("querySingleRelation failed, s.innerQueryRelationSingleModel error:%v", rvErr.Error())
 		err = rvErr
 		return
 	}
-
-	if rvModel != nil {
-		modelVal, modelErr := s.modelProvider.GetEntityValue(rvModel.Interface(vType.IsPtrType(), model.LiteView))
-		if modelErr != nil {
-			err = modelErr
-			log.Errorf("querySingleRelation failed, s.modelProvider.GetEntityValue error:%v", err.Error())
-			return
-		}
-
-		ret = modelVal
-		return
-	}
-
-	modelVal, modelErr := s.modelProvider.GetEntityValue(rModel.Interface(vType.IsPtrType(), model.LiteView))
-	if modelErr != nil {
-		err = modelErr
-		log.Errorf("querySingleRelation failed, s.modelProvider.GetEntityValue error:%v", err.Error())
-		return
-	}
-
-	ret = modelVal
 	return
 }
 
-func (s *QueryRunner) querySliceRelation(vModel model.Model, vField model.Field, rModel model.Model, deepLevel int) (ret model.Value, err *cd.Result) {
-	vType := vField.GetType()
-	vValue, _ := vType.Interface(nil)
-	valueList, valueErr := s.innerQueryRelationKeys(vModel, vField, rModel)
+func (s *QueryRunner) querySliceRelation(vModel model.Model, vField model.Field, deepLevel int) (err *cd.Error) {
+	valueList, valueErr := s.innerQueryRelationKeys(vModel, vField)
 	if valueErr != nil {
 		err = valueErr
-		if err.Fail() {
-			log.Errorf("querySliceRelation field:%s failed, s.innerQueryRelationKeys error:%sv", vField.GetName(), err.Error())
-		} else if err.Warn() {
-			log.Warnf("querySliceRelation field:%s failed, s.innerQueryRelationKeys error:%sv", vField.GetName(), err.Error())
-		}
+		log.Errorf("querySliceRelation field:%s failed, s.innerQueryRelationKeys error:%sv", vField.GetName(), err.Error())
 		return
 	}
 	valueSize := len(valueList)
 	if valueSize == 0 {
-		ret = vValue
 		return
 	}
 
-	rModelList, rModelErr := s.innerQueryRelationSliceModel(valueList, rModel, deepLevel)
+	rModelErr := s.innerQueryRelationSliceModel(valueList, vField, deepLevel)
 	if rModelErr != nil {
 		err = rModelErr
-		if err.Fail() {
-			log.Errorf("querySliceRelation field:%s failed, s.innerQueryRelationSliceModel error:%sv", vField.GetName(), err.Error())
-		} else if err.Warn() {
-			log.Warnf("querySliceRelation field:%s failed, s.innerQueryRelationSliceModel error:%sv", vField.GetName(), err.Error())
-		}
+		log.Errorf("querySliceRelation field:%s failed, s.innerQueryRelationSliceModel error:%sv", vField.GetName(), err.Error())
 		return
 	}
-
-	var vErr *cd.Result
-	elemType := vType.Elem()
-	for _, sv := range rModelList {
-		modelVal, modelErr := s.modelProvider.GetEntityValue(sv.Interface(elemType.IsPtrType(), model.LiteView))
-		if modelErr != nil {
-			err = modelErr
-			log.Errorf("querySliceRelation field:%s failed, s.modelProvider.GetEntityValue error:%sv", vField.GetName(), err.Error())
-			return
-		}
-
-		vValue, vErr = s.modelProvider.AppendSliceValue(vValue, modelVal)
-		if vErr != nil {
-			err = vErr
-			log.Errorf("querySliceRelation field:%s failed, s.modelProvider.AppendSliceValue error:%sv", vField.GetName(), err.Error())
-			return
-		}
-	}
-	ret = vValue
 	return
 }
 
-func (s *QueryRunner) innerQueryRelationKeys(vModel model.Model, vField model.Field, rModel model.Model) (ret resultItems, err *cd.Result) {
-	relationResult, relationErr := s.hBuilder.BuildQueryRelation(vModel, vField, rModel)
+func (s *QueryRunner) innerQueryRelationKeys(vModel model.Model, vField model.Field) (ret resultItems, err *cd.Error) {
+	relationResult, relationErr := s.hBuilder.BuildQueryRelation(vModel, vField)
 	if relationErr != nil {
 		err = relationErr
 		log.Errorf("innerQueryRelationKeys field:%s failed, hBuilder.BuildQueryRelation error:%v", vField.GetName(), err.Error())
@@ -309,14 +209,10 @@ func (s *QueryRunner) innerQueryRelationKeys(vModel model.Model, vField model.Fi
 		defer s.executor.Finish()
 
 		for s.executor.Next() {
-			itemValue, itemErr := s.hBuilder.BuildQueryRelationPlaceHolder(vModel, vField, rModel)
+			itemValue, itemErr := s.hBuilder.BuildQueryRelationPlaceHolder(vModel, vField)
 			if itemErr != nil {
 				err = itemErr
-				if err.Fail() {
-					log.Errorf("innerQueryRelationKeys field:%s failed, s.getModelPKFieldPlaceHolder error:%v", vField.GetName(), err.Error())
-				} else if err.Warn() {
-					log.Warnf("innerQueryRelationKeys field:%s failed, s.getModelPKFieldPlaceHolder error:%v", vField.GetName(), err.Error())
-				}
+				log.Errorf("innerQueryRelationKeys field:%s failed, s.getModelPKFieldPlaceHolder error:%v", vField.GetName(), err.Error())
 				return
 			}
 
@@ -325,7 +221,10 @@ func (s *QueryRunner) innerQueryRelationKeys(vModel model.Model, vField model.Fi
 				log.Errorf("innerQueryRelationKeys field:%s failed, s.executor.GetField error:%v", vField.GetName(), err.Error())
 				return
 			}
-			values = append(values, itemValue)
+
+			// 这里需要去除指针
+			rawVal := reflect.Indirect(reflect.ValueOf(itemValue)).Interface()
+			values = append(values, rawVal)
 		}
 	}()
 
@@ -337,24 +236,20 @@ func (s *QueryRunner) innerQueryRelationKeys(vModel model.Model, vField model.Fi
 	return
 }
 
-func (s *QueryRunner) innerQueryRelationSingleModel(id any, rModel model.Model, deepLevel int) (ret model.Model, err *cd.Result) {
-	pkField := rModel.GetPrimaryField()
-	fVal, fErr := s.modelProvider.DecodeValue(model.NewRawVal(id), pkField.GetType())
-	if fErr != nil {
-		err = fErr
-		log.Errorf("innerQueryRelationSingleModel failed, s.modelProvider.DecodeValue error:%v", err.Error())
+func (s *QueryRunner) innerQueryRelationSingleModel(id any, vField model.Field, deepLevel int) (err *cd.Error) {
+	vField.Reset()
+	rModel, rErr := s.modelProvider.GetTypeModel(vField.GetType())
+	if rErr != nil {
+		err = rErr
+		log.Errorf("innerQueryRelationSingleModel failed, s.modelProvider.GetTypeModel field:%s, id:%v, error:%v", vField.GetType().GetPkgKey(), id, err.Error())
 		return
 	}
-	rModel.SetFieldValue(pkField.GetName(), fVal)
 
-	vFilter, vErr := getModelFilter(rModel, s.modelProvider, model.LiteView)
+	rModel.SetPrimaryFieldValue(id)
+	vFilter, vErr := getModelFilter(rModel, s.modelProvider, s.modelCodec)
 	if vErr != nil {
 		err = vErr
-		if err.Fail() {
-			log.Errorf("innerQueryRelationSingleModel failed, getModelFilter error:%v", err.Error())
-		} else if err.Warn() {
-			log.Warnf("innerQueryRelationSingleModel failed, getModelFilter error:%v", err.Error())
-		}
+		log.Errorf("innerQueryRelationSingleModel failed, getModelFilter model:%s, id:%v, error:%v", rModel.GetPkgKey(), id, err.Error())
 		return
 	}
 
@@ -362,41 +257,45 @@ func (s *QueryRunner) innerQueryRelationSingleModel(id any, rModel model.Model, 
 	queryVal, queryErr := rQueryRunner.Query(vFilter)
 	if queryErr != nil {
 		err = queryErr
-		if err.Fail() {
-			log.Errorf("innerQueryRelationSingleModel failed, s.querySingle error:%v", err.Error())
-		} else if err.Warn() {
-			log.Warnf("innerQueryRelationSingleModel failed, s.querySingle error:%v", err.Error())
-		}
+		log.Errorf("innerQueryRelationSingleModel failed, s.querySingle model:%s, id:%v, error:%v", rModel.GetPkgKey(), id, err.Error())
+		return
+	}
+	if len(queryVal) > 1 {
+		errMsg := fmt.Sprintf("match more than one model, model:%s, id:%v", rModel.GetPkgKey(), id)
+		log.Warnf("innerQueryRelationSingleModel failed, errMsg:%s", errMsg)
+		err = cd.NewError(cd.Unexpected, errMsg)
 		return
 	}
 
 	if len(queryVal) > 0 {
-		ret = queryVal[0]
+		vField.SetValue(queryVal[0].Interface(vField.GetType().Elem().IsPtrType()))
+		return
+	}
+
+	if deepLevel < maxDeepLevel {
+		// 到这里说明未查询到数据，说明存在数据表之间数据不一致
+		// 这种情况下直接返回nil，后续要考虑进行脏数据检测
+		log.Warnf("query relation failed, miss relation data, model:%s, id:%v", rModel.GetPkgKey(), id)
 	}
 	return
 }
 
-func (s *QueryRunner) innerQueryRelationSliceModel(ids []any, rModel model.Model, deepLevel int) (ret []model.Model, err *cd.Result) {
-	sliceVal := []model.Model{}
+func (s *QueryRunner) innerQueryRelationSliceModel(ids []any, vField model.Field, deepLevel int) (err *cd.Error) {
+	// 这里主动重置，避免VFiled的旧数据干扰
+	vField.Reset()
 	for _, id := range ids {
-		svModel := rModel.Copy(false)
-		pkField := svModel.GetPrimaryField()
-		fVal, fErr := s.modelProvider.DecodeValue(model.NewRawVal(id), pkField.GetType())
-		if fErr != nil {
-			err = fErr
-			log.Errorf("innerQueryRelationSliceModel failed, s.modelProvider.DecodeValue error:%v", err.Error())
+		svModel, svErr := s.modelProvider.GetTypeModel(vField.GetType().Elem())
+		if svErr != nil {
+			err = svErr
+			log.Errorf("innerQueryRelationSliceModel failed, s.modelProvider.GetTypeModel error:%v", err.Error())
 			return
 		}
-		svModel.SetFieldValue(pkField.GetName(), fVal)
 
-		vFilter, vErr := getModelFilter(svModel, s.modelProvider, model.LiteView)
+		svModel.SetPrimaryFieldValue(id)
+		vFilter, vErr := getModelFilter(svModel, s.modelProvider, s.modelCodec)
 		if vErr != nil {
 			err = vErr
-			if err.Fail() {
-				log.Errorf("innerQueryRelationSliceModel failed, getModelFilter error:%v", err.Error())
-			} else if err.Warn() {
-				log.Warnf("innerQueryRelationSliceModel failed, getModelFilter error:%v", err.Error())
-			}
+			log.Errorf("innerQueryRelationSliceModel failed, getModelFilter error:%v", err.Error())
 			return
 		}
 
@@ -404,40 +303,33 @@ func (s *QueryRunner) innerQueryRelationSliceModel(ids []any, rModel model.Model
 		queryVal, queryErr := rQueryRunner.Query(vFilter)
 		if queryErr != nil {
 			err = queryErr
-			if err.Fail() {
-				log.Errorf("innerQueryRelationSingleModel failed, s.querySingle error:%v", err.Error())
-			} else if err.Warn() {
-				log.Warnf("innerQueryRelationSingleModel failed, s.querySingle error:%v", err.Error())
-			}
+			log.Errorf("innerQueryRelationSingleModel failed, s.querySingle error:%v", err.Error())
 			return
 		}
 
 		if len(queryVal) > 0 {
-			sliceVal = append(sliceVal, queryVal[0])
+			vField.AppendSliceValue(queryVal[0].Interface(vField.GetType().Elem().IsPtrType()))
 		}
 	}
 
-	ret = sliceVal
 	return
 }
 
-func (s *QueryRunner) Query(filter model.Filter) (ret []model.Model, err *cd.Result) {
+func (s *QueryRunner) Query(filter model.Filter) (ret []model.Model, err *cd.Error) {
 	queryValueList, queryValueErr := s.innerQuery(s.vModel, filter)
 	if queryValueErr != nil {
 		err = queryValueErr
-		if err.Fail() {
-			log.Errorf("Query failed, s.innerQuery error:%s", err.Error())
-		} else if err.Warn() {
-			log.Warnf("Query failed, s.innerQuery error:%s", err.Error())
-		}
+		log.Errorf("Query failed, s.innerQuery error:%s", err.Error())
 		return
 	}
+
 	queryCount := len(queryValueList)
 	if queryCount == 0 {
 		return
 	}
 	if !s.batchFilter && queryCount > 1 {
-		err = cd.NewResult(cd.UnExpected, fmt.Sprintf("match %d items", queryCount))
+		err = cd.NewError(cd.Unexpected, fmt.Sprintf("matched model:%s %d items value", s.vModel.GetPkgKey(), queryCount))
+		log.Warnf("Query failed, s.innerQuery warning:%s", err.Error())
 		return
 	}
 
@@ -446,11 +338,7 @@ func (s *QueryRunner) Query(filter model.Filter) (ret []model.Model, err *cd.Res
 		modelVal, modelErr := s.innerAssign(s.vModel, queryValueList[idx], 0)
 		if modelErr != nil {
 			err = modelErr
-			if err.Fail() {
-				log.Errorf("Query failed, s.innerAssign error:%s", err.Error())
-			} else if err.Warn() {
-				log.Warnf("Query failed, s.innerAssign error:%s", err.Error())
-			}
+			log.Errorf("Query failed, s.innerAssign error:%s", err.Error())
 			return
 		}
 
@@ -461,20 +349,18 @@ func (s *QueryRunner) Query(filter model.Filter) (ret []model.Model, err *cd.Res
 	return
 }
 
-func (s *impl) Query(vModel model.Model) (ret model.Model, err *cd.Result) {
+func (s *impl) Query(vModel model.Model) (ret model.Model, err *cd.Error) {
 	if vModel == nil {
-		err = cd.NewResult(cd.IllegalParam, "illegal model value")
+		err = cd.NewError(cd.IllegalParam, "illegal model value")
 		return
 	}
 
-	vFilter, vErr := getModelFilter(vModel, s.modelProvider, model.OriginView)
+	// 这里主动Copy一份出来，是为了避免在查询数据过程中对源数据产生了干扰
+	vModel = vModel.Copy(model.OriginView)
+	vFilter, vErr := getModelFilter(vModel, s.modelProvider, s.modelCodec)
 	if vErr != nil {
 		err = vErr
-		if err.Fail() {
-			log.Errorf("Query failed, s.getModelFilter error:%v", err.Error())
-		} else if err.Warn() {
-			log.Warnf("Query failed, s.getModelFilter error:%v", err.Error())
-		}
+		log.Errorf("Query failed, s.getModelFilter error:%v", err.Error())
 		return
 	}
 
@@ -482,11 +368,7 @@ func (s *impl) Query(vModel model.Model) (ret model.Model, err *cd.Result) {
 	queryVal, queryErr := vQueryRunner.Query(vFilter)
 	if queryErr != nil {
 		err = queryErr
-		if err.Fail() {
-			log.Errorf("Query failed, vQueryRunner.Query error:%v", err.Error())
-		} else if err.Warn() {
-			log.Warnf("Query failed, vQueryRunner.Query error:%v", err.Error())
-		}
+		log.Errorf("Query failed, vQueryRunner.Query error:%v", err.Error())
 		return
 	}
 	if len(queryVal) != 0 {
@@ -494,6 +376,6 @@ func (s *impl) Query(vModel model.Model) (ret model.Model, err *cd.Result) {
 		return
 	}
 
-	err = cd.NewResult(cd.NoExist, fmt.Sprintf("query model failed, pkgKey:%s", vModel.GetPkgKey()))
+	err = cd.NewError(cd.NotFound, fmt.Sprintf("query model failed, model:%s", vModel.GetPkgKey()))
 	return
 }
