@@ -1,99 +1,178 @@
 package local
 
 import (
+	"fmt"
 	"reflect"
 
+	cd "github.com/muidea/magicCommon/def"
+	"github.com/muidea/magicCommon/foundation/log"
 	"github.com/muidea/magicOrm/model"
-	"github.com/muidea/magicOrm/provider/util"
 )
 
 type ValueImpl struct {
 	value reflect.Value
 }
 
-var NilValue = ValueImpl{}
-
 func NewValue(val reflect.Value) (ret *ValueImpl) {
 	ret = &ValueImpl{value: val}
 	return
 }
 
-func (s *ValueImpl) IsValid() (ret bool) {
+func (s *ValueImpl) IsValid() bool {
 	if !s.value.IsValid() {
 		return false
 	}
-
-	rawVal := reflect.Indirect(s.value)
-	switch rawVal.Kind() {
-	case reflect.Slice, reflect.Map:
-		return !rawVal.IsNil()
-	default:
+	if s.value.Kind() == reflect.Ptr {
+		if s.value.IsNil() {
+			return false
+		}
 	}
-	ret = rawVal.IsValid()
-	return
+	if s.value.Kind() == reflect.Slice || s.value.Kind() == reflect.Map {
+		if s.value.IsNil() {
+			return false
+		}
+	}
+
+	return true
 }
 
-func (s *ValueImpl) IsZero() (ret bool) {
-	if util.IsNil(s.value) {
-		ret = true
+func (s *ValueImpl) IsZero() bool {
+	if !s.value.IsValid() {
+		return true
+	}
+	if s.value.Kind() == reflect.Ptr {
+		if s.value.IsNil() {
+			return true
+		}
+		return s.value.Elem().IsZero()
+	}
+	if s.value.Kind() == reflect.Slice || s.value.Kind() == reflect.Map {
+		if s.value.IsNil() {
+			return true
+		}
+		return s.value.Len() == 0
+	}
+
+	return s.value.IsZero()
+}
+
+func (s *ValueImpl) Set(val any) (err *cd.Error) {
+	if !s.value.CanSet() {
+		err = cd.NewError(cd.Unexpected, "Set failed, value is not settable")
+		log.Warnf("Set failed, value is not settable")
+		return
+	}
+	if !s.value.IsValid() {
+		log.Errorf("Set failed, value is not valid, s.value canSet:%+v", s.value.CanSet())
 		return
 	}
 
-	ret = util.IsZero(s.value)
-	return
-}
+	rVal := reflect.ValueOf(val)
+	isPtr := s.value.Kind() == reflect.Ptr
+	if !isPtr {
+		if rVal.Type() != s.value.Type() {
+			err = cd.NewError(cd.Unexpected, "Set failed, value type is not match")
+			log.Warnf("Set failed, value type is not match, data type:%+v, value type:%+v", rVal.Type(), s.value.Type())
+			return
+		}
 
-func (s *ValueImpl) Set(val any) {
-	s.value = val.(reflect.Value)
+		s.value.Set(rVal)
+		return
+	}
+
+	rVal = reflect.Indirect(rVal)
+	if rVal.Type() != s.value.Type().Elem() {
+		err = cd.NewError(cd.Unexpected, "Set failed, value type is not match")
+		log.Warnf("Set failed, value type is not match")
+		return
+	}
+
+	reallyValPtr := reflect.New(s.value.Type().Elem())
+	reallyValPtr.Elem().Set(rVal)
+	s.value.Set(reallyValPtr)
+	return
 }
 
 func (s *ValueImpl) Get() any {
-	return s.value
+	return s.value.Interface()
 }
 
-func (s *ValueImpl) Addr() model.Value {
-	if !s.value.CanAddr() {
-		panic("illegal value, can't addr")
+// UnpackValue expands the contained value into individual elements.
+// For slices, it returns each element as separate reflect.Value entries.
+// For non-slice values, returns a single-element slice containing the value.
+func (s *ValueImpl) UnpackValue() (ret []model.Value) {
+	ret = []model.Value{}
+	realVal := reflect.Indirect(s.value)
+	if realVal.Kind() == reflect.Slice {
+		for idx := 0; idx < realVal.Len(); idx++ {
+			ret = append(ret, NewValue(realVal.Index(idx)))
+		}
+	} else {
+		ret = append(ret, NewValue(s.value))
 	}
-
-	impl := &ValueImpl{value: s.value.Addr()}
-	return impl
+	return
 }
 
-func (s *ValueImpl) Interface() model.RawVal {
-	if util.IsNil(s.value) {
-		return nil
-	}
+// Append appends the given value to the slice value.
+func (s *ValueImpl) Append(val reflect.Value) (err *cd.Error) {
+	defer func() {
+		if errInfo := recover(); errInfo != nil {
+			err = cd.NewError(cd.Unexpected, fmt.Sprintf("%v", errInfo))
+		}
+	}()
 
-	return model.NewRawVal(s.value.Interface())
-}
+	isPtr := s.value.Kind() == reflect.Ptr
+	if !isPtr {
+		//if !s.IsValid() {
+		//	initSlice := reflect.MakeSlice(s.value.Type(), 0, 0)
+		//	s.value.Set(initSlice)
+		//}
 
-func (s *ValueImpl) IsBasic() bool {
-	if util.IsNil(s.value) {
-		return false
-	}
-
-	rType := s.value.Type()
-	if rType.Kind() == reflect.Ptr {
-		rType = rType.Elem()
-	}
-	if s.value.Kind() == reflect.Interface {
-		rType = s.value.Elem().Type()
-	}
-	if util.IsSlice(rType) {
-		rType = rType.Elem()
-	}
-
-	return !util.IsStruct(rType)
-}
-
-func (s *ValueImpl) Copy() (ret *ValueImpl) {
-	if !util.IsNil(s.value) {
-		ret = &ValueImpl{value: reflect.New(s.value.Type()).Elem()}
-		ret.value.Set(s.value)
+		s.value.Set(reflect.Append(s.value, val))
 		return
 	}
 
-	ret = &ValueImpl{}
+	if !s.value.IsValid() || s.value.IsZero() {
+		initSlice := reflect.MakeSlice(s.value.Type().Elem(), 0, 0)
+		initSlicePtr := reflect.New(s.value.Type().Elem())
+		initSlicePtr.Elem().Set(initSlice)
+		s.value.Set(initSlicePtr)
+	}
+
+	reallySlice := reflect.Indirect(s.value)
+	reallySlice = reflect.Append(reallySlice, val)
+	reallySlicePtr := reflect.New(reallySlice.Type())
+	reallySlicePtr.Elem().Set(reallySlice)
+	s.value.Set(reallySlicePtr)
 	return
+}
+
+func (s *ValueImpl) reset(needInit bool) {
+	if !s.value.IsValid() {
+		return
+	}
+
+	if s.value.Kind() == reflect.Ptr {
+		if needInit {
+			s.value.Set(reflect.New(s.value.Type().Elem()))
+		} else {
+			s.value.Set(reflect.Zero(s.value.Type()))
+		}
+		return
+	}
+
+	if s.value.Kind() == reflect.Slice || s.value.Kind() == reflect.Map {
+		if needInit {
+			if s.value.Kind() == reflect.Slice {
+				s.value.Set(reflect.MakeSlice(s.value.Type(), 0, 0))
+			} else {
+				s.value.Set(reflect.MakeMap(s.value.Type()))
+			}
+		} else {
+			s.value.Set(reflect.Zero(s.value.Type()))
+		}
+		return
+	}
+
+	s.value.SetZero()
 }
