@@ -3,40 +3,39 @@ package remote
 import (
 	"encoding/json"
 	"fmt"
-	"path"
 
 	cd "github.com/muidea/magicCommon/def"
 	"github.com/muidea/magicCommon/foundation/log"
-	om "github.com/muidea/magicOrm/model"
-	pu "github.com/muidea/magicOrm/provider/util"
+	"github.com/muidea/magicOrm/model"
+	"github.com/muidea/magicOrm/utils"
 )
 
 type filterItem struct {
-	oprCode om.OprCode
+	oprCode model.OprCode
 	value   *ValueImpl
 }
 
-func (s *filterItem) OprCode() om.OprCode {
+func (s *filterItem) OprCode() model.OprCode {
 	return s.oprCode
 }
 
-func (s *filterItem) OprValue() om.Value {
+func (s *filterItem) OprValue() model.Value {
 	return s.value
 }
 
 type ObjectFilter struct {
-	Name           string         `json:"name"`
-	PkgPath        string         `json:"pkgPath"`
-	EqualFilter    []*FieldValue  `json:"equal"`
-	NotEqualFilter []*FieldValue  `json:"noEqual"`
-	BelowFilter    []*FieldValue  `json:"below"`
-	AboveFilter    []*FieldValue  `json:"above"`
-	InFilter       []*FieldValue  `json:"in"`
-	NotInFilter    []*FieldValue  `json:"notIn"`
-	LikeFilter     []*FieldValue  `json:"like"`
-	MaskValue      *ObjectValue   `json:"maskValue"`
-	PageFilter     *pu.Pagination `json:"page"`
-	SortFilter     *pu.SortFilter `json:"sort"`
+	Name           string            `json:"name"`
+	PkgPath        string            `json:"pkgPath"`
+	EqualFilter    []*FieldValue     `json:"equal"`
+	NotEqualFilter []*FieldValue     `json:"noEqual"`
+	BelowFilter    []*FieldValue     `json:"below"`
+	AboveFilter    []*FieldValue     `json:"above"`
+	InFilter       []*FieldValue     `json:"in"`
+	NotInFilter    []*FieldValue     `json:"notIn"`
+	LikeFilter     []*FieldValue     `json:"like"`
+	MaskValue      *ObjectValue      `json:"maskValue"`
+	PageFilter     *utils.Pagination `json:"page"`
+	SortFilter     *utils.SortFilter `json:"sort"`
 
 	bindObject *Object
 }
@@ -62,10 +61,6 @@ func (s *ObjectFilter) GetName() string {
 
 func (s *ObjectFilter) GetPkgPath() string {
 	return s.PkgPath
-}
-
-func (s *ObjectFilter) GetPkgKey() string {
-	return path.Join(s.GetPkgPath(), s.GetName())
 }
 
 func (s *ObjectFilter) GetString(key string) (ret string, ok bool) {
@@ -96,7 +91,7 @@ func (s *ObjectFilter) GetInt(key string) (ret int, ok bool) {
 	return
 }
 
-func (s *ObjectFilter) Equal(key string, val any) (err *cd.Result) {
+func (s *ObjectFilter) Equal(key string, val any) (err *cd.Error) {
 	switch val.(type) {
 	case bool,
 		int8, int16, int32, int, int64,
@@ -112,13 +107,13 @@ func (s *ObjectFilter) Equal(key string, val any) (err *cd.Result) {
 		}
 		s.EqualFilter = append(s.EqualFilter, item)
 	default:
-		err = cd.NewResult(cd.UnExpected, fmt.Sprintf("equal failed, illegal value, key:%v, val:%v", key, val))
+		err = cd.NewError(cd.Unexpected, fmt.Sprintf("equal failed, illegal value, key:%v, val:%v", key, val))
 	}
 
 	return
 }
 
-func (s *ObjectFilter) NotEqual(key string, val any) (err *cd.Result) {
+func (s *ObjectFilter) NotEqual(key string, val any) (err *cd.Error) {
 	switch val.(type) {
 	case bool,
 		int8, int16, int32, int, int64,
@@ -134,13 +129,13 @@ func (s *ObjectFilter) NotEqual(key string, val any) (err *cd.Result) {
 		}
 		s.NotEqualFilter = append(s.NotEqualFilter, item)
 	default:
-		err = cd.NewResult(cd.UnExpected, fmt.Sprintf("not equal failed, illegal value, key:%v, val:%v", key, val))
+		err = cd.NewError(cd.Unexpected, fmt.Sprintf("not equal failed, illegal value, key:%v, val:%v", key, val))
 	}
 
 	return
 }
 
-func (s *ObjectFilter) Below(key string, val any) (err *cd.Result) {
+func (s *ObjectFilter) Below(key string, val any) (err *cd.Error) {
 	switch val.(type) {
 	case int8, int16, int32, int, int64,
 		uint8, uint16, uint32, uint, uint64,
@@ -148,12 +143,12 @@ func (s *ObjectFilter) Below(key string, val any) (err *cd.Result) {
 		item := &FieldValue{Name: key, Value: val}
 		s.BelowFilter = append(s.BelowFilter, item)
 	default:
-		err = cd.NewResult(cd.UnExpected, fmt.Sprintf("below failed, illegal value, key:%v, val:%v", key, val))
+		err = cd.NewError(cd.Unexpected, fmt.Sprintf("below failed, illegal value, key:%v, val:%v", key, val))
 	}
 	return
 }
 
-func (s *ObjectFilter) Above(key string, val any) (err *cd.Result) {
+func (s *ObjectFilter) Above(key string, val any) (err *cd.Error) {
 	switch val.(type) {
 	case int8, int16, int32, int, int64,
 		uint8, uint16, uint32, uint, uint64,
@@ -161,12 +156,12 @@ func (s *ObjectFilter) Above(key string, val any) (err *cd.Result) {
 		item := &FieldValue{Name: key, Value: val}
 		s.AboveFilter = append(s.AboveFilter, item)
 	default:
-		err = cd.NewResult(cd.UnExpected, fmt.Sprintf("above failed, illegal value, key:%v, val:%v", key, val))
+		err = cd.NewError(cd.Unexpected, fmt.Sprintf("above failed, illegal value, key:%v, val:%v", key, val))
 	}
 	return
 }
 
-func (s *ObjectFilter) In(key string, val any) (err *cd.Result) {
+func (s *ObjectFilter) In(key string, val any) (err *cd.Error) {
 	switch val.(type) {
 	case []bool,
 		[]int8, []int16, []int32, []int, []int64,
@@ -183,13 +178,13 @@ func (s *ObjectFilter) In(key string, val any) (err *cd.Result) {
 		}
 		s.InFilter = append(s.InFilter, item)
 	default:
-		err = cd.NewResult(cd.UnExpected, fmt.Sprintf("in failed, illegal value, key:%v, val:%v", key, val))
+		err = cd.NewError(cd.Unexpected, fmt.Sprintf("in failed, illegal value, key:%v, val:%v", key, val))
 	}
 
 	return
 }
 
-func (s *ObjectFilter) NotIn(key string, val any) (err *cd.Result) {
+func (s *ObjectFilter) NotIn(key string, val any) (err *cd.Error) {
 	switch val.(type) {
 	case []bool,
 		[]int8, []int16, []int32, []int, []int64,
@@ -206,41 +201,41 @@ func (s *ObjectFilter) NotIn(key string, val any) (err *cd.Result) {
 		}
 		s.NotInFilter = append(s.NotInFilter, item)
 	default:
-		err = cd.NewResult(cd.UnExpected, fmt.Sprintf("not in failed, illegal value, key:%v, val:%v", key, val))
+		err = cd.NewError(cd.Unexpected, fmt.Sprintf("not in failed, illegal value, key:%v, val:%v", key, val))
 	}
 
 	return
 }
 
-func (s *ObjectFilter) Like(key string, val any) (err *cd.Result) {
+func (s *ObjectFilter) Like(key string, val any) (err *cd.Error) {
 	switch val.(type) {
 	case string:
 		item := &FieldValue{Name: key, Value: val}
 		s.LikeFilter = append(s.LikeFilter, item)
 	default:
-		err = cd.NewResult(cd.UnExpected, fmt.Sprintf("like failed, illegal value, key:%v, val:%v", key, val))
+		err = cd.NewError(cd.Unexpected, fmt.Sprintf("like failed, illegal value, key:%v, val:%v", key, val))
 	}
 
 	return
 }
 
 func (s *ObjectFilter) Pagination(pageNum, pageSize int) {
-	s.PageFilter = &pu.Pagination{
+	s.PageFilter = &utils.Pagination{
 		PageNum:  pageNum,
 		PageSize: pageSize,
 	}
 }
 
 func (s *ObjectFilter) Sort(fieldName string, ascFlag bool) {
-	s.SortFilter = &pu.SortFilter{
+	s.SortFilter = &utils.SortFilter{
 		FieldName: fieldName,
 		AscFlag:   ascFlag,
 	}
 }
 
-func (s *ObjectFilter) ValueMask(val any) (err *cd.Result) {
+func (s *ObjectFilter) ValueMask(val any) (err *cd.Error) {
 	if val == nil {
-		err = cd.NewResult(cd.UnExpected, "illegal mask value")
+		err = cd.NewError(cd.Unexpected, "illegal mask value")
 		return
 	}
 
@@ -255,7 +250,7 @@ func (s *ObjectFilter) ValueMask(val any) (err *cd.Result) {
 	case json.RawMessage:
 		objectValuePtr, err = DecodeObjectValue(v)
 	default:
-		err = cd.NewResult(cd.UnExpected, "illegal mask value")
+		err = cd.NewError(cd.Unexpected, "illegal mask value")
 	}
 
 	if err != nil {
@@ -267,23 +262,17 @@ func (s *ObjectFilter) ValueMask(val any) (err *cd.Result) {
 		return
 	}
 
-	if s.bindObject.GetPkgKey() != objectValuePtr.GetPkgKey() {
-		err = cd.NewResult(cd.UnExpected, fmt.Sprintf("mismatch mask value, bindPkgKey:%v, maskPkgKey:%v", s.bindObject.GetPkgKey(), objectValuePtr.GetPkgKey()))
-		log.Errorf("ValueMask failed, err:%v", err.Error())
-		return
-	}
-
 	s.MaskValue = objectValuePtr
 	return
 }
 
-func (s *ObjectFilter) GetFilterItem(key string) om.FilterItem {
+func (s *ObjectFilter) GetFilterItem(key string) model.FilterItem {
 	itemVal, itemErr := s.getFilterValue(key, s.EqualFilter)
 	if itemErr != nil {
 		return nil
 	}
 	if itemVal != nil {
-		return &filterItem{oprCode: om.EqualOpr, value: NewValue(itemVal.Get())}
+		return &filterItem{oprCode: model.EqualOpr, value: NewValue(itemVal.Get())}
 	}
 
 	itemVal, itemErr = s.getFilterValue(key, s.NotEqualFilter)
@@ -291,7 +280,7 @@ func (s *ObjectFilter) GetFilterItem(key string) om.FilterItem {
 		return nil
 	}
 	if itemVal != nil {
-		return &filterItem{oprCode: om.NotEqualOpr, value: NewValue(itemVal.Get())}
+		return &filterItem{oprCode: model.NotEqualOpr, value: NewValue(itemVal.Get())}
 	}
 
 	itemVal, itemErr = s.getFilterValue(key, s.BelowFilter)
@@ -299,7 +288,7 @@ func (s *ObjectFilter) GetFilterItem(key string) om.FilterItem {
 		return nil
 	}
 	if itemVal != nil {
-		return &filterItem{oprCode: om.BelowOpr, value: NewValue(itemVal.Get())}
+		return &filterItem{oprCode: model.BelowOpr, value: NewValue(itemVal.Get())}
 	}
 
 	itemVal, itemErr = s.getFilterValue(key, s.AboveFilter)
@@ -307,7 +296,7 @@ func (s *ObjectFilter) GetFilterItem(key string) om.FilterItem {
 		return nil
 	}
 	if itemVal != nil {
-		return &filterItem{oprCode: om.AboveOpr, value: NewValue(itemVal.Get())}
+		return &filterItem{oprCode: model.AboveOpr, value: NewValue(itemVal.Get())}
 	}
 
 	itemVal, itemErr = s.getFilterValue(key, s.InFilter)
@@ -315,7 +304,7 @@ func (s *ObjectFilter) GetFilterItem(key string) om.FilterItem {
 		return nil
 	}
 	if itemVal != nil {
-		return &filterItem{oprCode: om.InOpr, value: NewValue(itemVal.Get())}
+		return &filterItem{oprCode: model.InOpr, value: NewValue(itemVal.Get())}
 	}
 
 	itemVal, itemErr = s.getFilterValue(key, s.NotInFilter)
@@ -323,7 +312,7 @@ func (s *ObjectFilter) GetFilterItem(key string) om.FilterItem {
 		return nil
 	}
 	if itemVal != nil {
-		return &filterItem{oprCode: om.NotInOpr, value: NewValue(itemVal.Get())}
+		return &filterItem{oprCode: model.NotInOpr, value: NewValue(itemVal.Get())}
 	}
 
 	itemVal, itemErr = s.getFilterValue(key, s.LikeFilter)
@@ -331,13 +320,13 @@ func (s *ObjectFilter) GetFilterItem(key string) om.FilterItem {
 		return nil
 	}
 	if itemVal != nil {
-		return &filterItem{oprCode: om.LikeOpr, value: NewValue(itemVal.Get())}
+		return &filterItem{oprCode: model.LikeOpr, value: NewValue(itemVal.Get())}
 	}
 
 	return nil
 }
 
-func (s *ObjectFilter) getFilterValue(key string, items []*FieldValue) (ret *FieldValue, err *cd.Result) {
+func (s *ObjectFilter) getFilterValue(key string, items []*FieldValue) (ret *FieldValue, err *cd.Error) {
 	for _, val := range items {
 		if key == val.Name {
 			ret = val
@@ -351,7 +340,7 @@ func (s *ObjectFilter) getFilterValue(key string, items []*FieldValue) (ret *Fie
 	return
 }
 
-func (s *ObjectFilter) Paginationer() om.Paginationer {
+func (s *ObjectFilter) Paginationer() model.Paginationer {
 	if s.PageFilter == nil {
 		return nil
 	}
@@ -359,7 +348,7 @@ func (s *ObjectFilter) Paginationer() om.Paginationer {
 	return s.PageFilter
 }
 
-func (s *ObjectFilter) Sorter() om.Sorter {
+func (s *ObjectFilter) Sorter() model.Sorter {
 	if s.SortFilter == nil {
 		return nil
 	}
@@ -367,13 +356,13 @@ func (s *ObjectFilter) Sorter() om.Sorter {
 	return s.SortFilter
 }
 
-func (s *ObjectFilter) MaskModel() om.Model {
+func (s *ObjectFilter) MaskModel() model.Model {
 	maskObject := s.bindObject
 	if s.MaskValue != nil {
 		for _, val := range s.MaskValue.Fields {
-			maskObject.SetFieldValue(val.Name, val.GetValue())
+			maskObject.SetFieldValue(val.Name, val.GetValue().Get())
 		}
 	}
 
-	return maskObject.Copy(false)
+	return maskObject.Copy(model.OriginView)
 }
