@@ -1,5 +1,10 @@
 package test
 
+import (
+	cd "github.com/muidea/magicCommon/def"
+	"github.com/muidea/magicOrm/provider/remote"
+)
+
 type SKUInfo struct {
 	SKU         string   `json:"sku" orm:"sku key" view:"detail,lite"`
 	Description string   `json:"description" orm:"description" view:"detail,lite"`
@@ -8,12 +13,12 @@ type SKUInfo struct {
 }
 
 type Product struct {
-	ID          int64      `json:"id" orm:"id key auto" view:"detail,lite"`
-	Name        string     `json:"name" orm:"name" view:"detail,lite"`
-	Description string     `json:"description" orm:"description" view:"detail,lite"`
-	SKUInfo     []*SKUInfo `json:"skuInfo" orm:"skuInfo" view:"detail,lite"`
-	Image       []string   `json:"image" orm:"image" view:"detail,lite"`
-	Expire      int        `json:"expire" orm:"expire" view:"detail,lite"`
+	ID          int64     `json:"id" orm:"id key auto" view:"detail,lite"`
+	Name        string    `json:"name" orm:"name" view:"detail,lite"`
+	Description string    `json:"description" orm:"description" view:"detail,lite"`
+	SKUInfo     []SKUInfo `json:"skuInfo" orm:"skuInfo" view:"detail,lite"`
+	Image       []string  `json:"image" orm:"image" view:"detail,lite"`
+	Expire      int       `json:"expire" orm:"expire" view:"detail,lite"`
 }
 
 type Store struct {
@@ -36,4 +41,58 @@ type StockIn struct {
 	GoodsInfo   []GoodsInfo `json:"goodsInfo" view:"detail,lite"`
 	Description string      `json:"description" orm:"description" view:"detail,lite"`
 	Store       *Store      `json:"store" orm:"store" view:"detail,lite"`
+}
+
+func getLocalProduct() *Product {
+	return &Product{
+		Name:        "product",
+		Description: "product description",
+		SKUInfo: []SKUInfo{
+			{
+				SKU:         "sku001",
+				Description: "sku001 description",
+				Image:       []string{"image"},
+				Namespace:   "xyz",
+			},
+			{
+				SKU:         "sku002",
+				Description: "sku002 description",
+				Image:       []string{"image"},
+				Namespace:   "xyz",
+			},
+		},
+		Image:  []string{"image"},
+		Expire: 1,
+	}
+}
+
+func getLocalStore() *Store {
+	return &Store{
+		Code: "store001",
+		Name: "store001",
+	}
+}
+
+func getLocalStockIn(productPtr *Product, storePtr *Store) *StockIn {
+	return &StockIn{
+		SN:          "stockin001",
+		GoodsInfo:   []GoodsInfo{{SKU: "sku001", Product: productPtr, Count: 1, Price: 1}},
+		Description: "stockin001 description",
+		Store:       storePtr,
+	}
+}
+
+func getRemoteProduct(productPtr *Product) (ret *remote.ObjectValue, err *cd.Error) {
+	ret, err = getObjectValue(productPtr)
+	return
+}
+
+func getRemoteStore(storePtr *Store) (ret *remote.ObjectValue, err *cd.Error) {
+	ret, err = getObjectValue(storePtr)
+	return
+}
+
+func getRemoteStockIn(stockInPtr *StockIn) (ret *remote.ObjectValue, err *cd.Error) {
+	ret, err = getObjectValue(stockInPtr)
+	return
 }
